rename statusLinks
