+++ readURL.js
@@ -1,19 +1,16 @@
-
 const paths = require('./pathsFind.js');
 const index = require('./index.js');
 const fetch = require('node-fetch');
-const testFetch = require('./testFetch.js');
+const statusLinks = require('./statusLinks.js');
 
 
 const readLinkURL = (data) => {
 
   const regexLink = /(https?:\/\/[^\)\s ]+)/g;
   const regexMatch = data.match(regexLink);
-  testFetch.statusDate(regexMatch);
-  console.log('match',regexMatch);
+  statusLinks.statusDate(regexMatch);
+  console.log(regexMatch);
 
 };
 
 module.exports.readLinkURL = readLinkURL;
-
-
